backend/requirements.txt
fastapi
uvicorn
psycopg2-binary[extra]
sqlalchemy 
python-dotenv
pydantic-settings
app
python-multipart
fastapi[all]